Moved select catalog from overview to about page.
