+++ src/routes/About.svelte
@@ -17,10 +17,12 @@
   import ExpandCollapseAll from "../components/ExpandCollapseAll.svelte";
   import { honourFragmentIdLinks } from "../utils/honourFragmentIdLinks.js";
   import { reportFilename } from "../utils/reportFilename.js";
-  import { getCatalog } from "../utils/getCatalogs.js";
+  import { getCatalog, getListOfCatalogs } from "../utils/getCatalogs.js";
+  import { updateEvaluation } from "../utils/updateEvaluation.js";
 
   const location = useLocation();
   let catalog = getCatalog($evaluation.catalog);
+  let catalogChoices = getListOfCatalogs();
 
   onMount(() => {
     currentPage.update(currentPage => "About");
@@ -81,6 +83,16 @@
     }
   }
 
+  function updateCatalog(e) {
+    if (
+      window.confirm(
+        "This may remove criteria that are not in the selected catalog. Are you sure that's what you'd like to do?"
+      )
+    ) {
+      updateEvaluation(e.target.value, $evaluation);
+    }
+  }
+
   $: versionPrefix = reportFilename($evaluation, false);
 </script>
 
@@ -106,6 +118,27 @@
   />
 
 <ExpandCollapseAll />
+
+<details open>
+  <summary>
+    <HeaderWithAnchor id="select-catalog" level=2>Select catalog</HeaderWithAnchor>
+  </summary>
+  <p>{helpText["catalog"]["intro"]}</p>
+  {#each catalogChoices as catalogChoice}
+    <div class="field">
+      <label>
+        <input
+          type="radio"
+          value={catalogChoice.catalog}
+          bind:group="{$evaluation['catalog']}"
+          id="evaluation-catalog-{catalogChoice.catalog}"
+          on:change={updateCatalog} />
+
+        {catalogChoice.title}
+      </label>
+    </div>
+  {/each}
+</details>
 
 <details open>
   <summary>
