+++ src/routes/Overview.svelte
@@ -9,28 +9,16 @@
   import HeaderWithAnchor from "../components/HeaderWithAnchor.svelte";
   import { honourFragmentIdLinks } from "../utils/honourFragmentIdLinks.js";
   import { evaluation } from "../stores/evaluation.js";
-  import { getCatalog, getListOfCatalogs } from "../utils/getCatalogs.js";
-  import { updateEvaluation } from "../utils/updateEvaluation.js";
+  import { getCatalog } from "../utils/getCatalogs.js";
 
   const location = useLocation();
   let catalog = getCatalog($evaluation.catalog);
-  let catalogChoices = getListOfCatalogs();
 
   onMount(() => {
     currentPage.update((currentPage) => "Overview");
 
     honourFragmentIdLinks($location);
   });
-
-  function updateCatalog(e) {
-    if (
-      window.confirm(
-        "This may remove criteria that are not in the selected catalog. Are you sure that's what you'd like to do?"
-      )
-    ) {
-      updateEvaluation(e.target.value, $evaluation);
-    }
-  }
 </script>
 
 <svelte:head>
@@ -69,29 +57,6 @@
 </ul>
 
 <ExpandCollapseAll />
-
-<details open>
-  <summary>
-    <HeaderWithAnchor id="select-catalog" level=2>Select catalog</HeaderWithAnchor>
-  </summary>
-  <p>
-    Select which catalog you will be using for the OpenACR.
-  </p>
-  {#each catalogChoices as catalogChoice}
-    <div class="field">
-      <label>
-        <input
-          type="radio"
-          value={catalogChoice.catalog}
-          bind:group="{$evaluation['catalog']}"
-          id="evaluation-catalog-{catalogChoice.catalog}"
-          on:change={updateCatalog} />
-
-        {catalogChoice.title}
-      </label>
-    </div>
-  {/each}
-</details>
 
 <details>
   <summary>
